backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
social-auth-app-django>=5.4,<6.0
requests>=2.31,<3.0
Pillow>=10.0,<12.0
psycopg2-binary>=2.9,<3.0
django-storages[s3]>=1.14,<2.0
boto3>=1.34,<2.0
